#Added testing for constraints and gson views

+++ grails-app/controllers/restfulwebservice/WorkbookController.groovy
@@ -1,5 +1,6 @@
 package restfulwebservice
 
+import grails.converters.JSON
 import grails.gorm.transactions.Transactional
 
 
@@ -23,41 +24,62 @@ class WorkbookController {
     static responseFormats=['json','xml']
     def index()
     {
-        println('success index')
-        respond workbookService.list()
+        if(workbookService.list().size() <= 0)
+        {
+            response.status = 404
+            render([error:'workbooks not found'] as JSON)
+        }
+        else
+        {
+            println('success index')
+            respond workbookService.list()
+
+        }
 
 
     }
 
     def show(Workbook workbook) {
-        println('success show')
-        respond workbook
+        if(workbook == null)
+        {
+            response.status = 404
+            render([error:'workbook not found'] as JSON)
+        }
+        else
+        {
+            println('success show')
+            respond workbook
+        }
+
     }
 
     def delete (Workbook workbook){
-        println('success delete')
-        workbookService.delete(workbook)
-        withFormat {
-            html {
-//                    flash.message = message(code: 'default.created.message', args: [message(code: 'book.label', default: 'Book'), book.id])
-                redirect workbook
-            }
-//                '*' { render status: CREATED }
+        if(workbook == null)
+        {
+            response.status = 404
+            render([error:'workbook not found'] as JSON)
         }
+        else
+        {
+            println('success delete')
+            workbookService.delete(workbook)
+        }
+
     }
     @Transactional
     def save(Workbook workbook){
 //        println(workbook.dateOfBirth)
+        println(response)
         println(workbook.workplaces)
 
-
-        if(workbook.validate())
+         if(workbook.validate())
         {
 //            respond workbook.errors,view:'/workbook/create'
 //            respond workbook
             workbookService.save(workbook)
      }
         else{
+
             respond workbook.org_grails_datastore_gorm_GormValidateable__errors
             println workbook.org_grails_datastore_gorm_GormValidateable__errors
